simplify canvasUtil
